[Handshake] Memory operations rework
This commit makes several changes to memory-related operations used by
Dynamatic.

The MemoryControllerOp operands are reworked so that they hold complete
information about the per-block sequences of load and store accesses
interacting with the memory. This is achieved thanks to a custom MLIR
attribute stored as an operand to the operation. A verifier is also
added to the operation to check for consistency between the operation's
inputs/outputs and the specified memory accesses. Finally, the way this
operation connects to the rest of the circuit is slightly modified so
that it doesn't need to receive a control value fed through a constant
for blocks that do not contain any store operation.

The LoadOp and StoreOp operations are modified to only support one SSA
value for indexing (instead of potentially multiple indices), which is
the only use case we care about. This simplifies their interface as
well as the indexing logic in MemoryControllerOp's. Custom parser and
printer functions for these operations are also deleted in favor of a
TableGen assembly format. Finally, return type inference is added for
both these operations.