[Handshake] Support for buffering properties
This commit introduces a Handshake attribute and an accompanying
data-structure to hold channel buffering properties that will be used in
the buffer placement pass to control the nature and positions of
buffers in our elastic circuits.

Buffering properties may be attached to any channel (i.e., SSA value)
through its producer/defining operation. Multiple set of buffering
properties may thus be attached to an operation if it has multiple
results.