[Handshake] Get bitwidths of memory controller
This commit adds an helper method on MemoryControllerOp that allows to
easily get the bitwidths of control values (if present), address values,
and data values. This will be immediately useful for the
HandshakeToNetlist conversion pass.

The way memory controllers represent and give access to their associated
memory accesses needs a serious revamp, but that is for a later time.