[ChannelBufferProps] Name changes and bug fix
This commit changes the naming conventions used in ChannelBufferProps so
that it refers to "non-transparent" slots as "opaque" slots instead,
making it more consistent with legacy Dynamatic and literature
conventions.

The commit also fixes a bug in the two "subtraction" (not
"substraction", as previously spelled) methods of the struct where a
subtraction between two unsigned would lead to an underflow in some
situtations.