[StandardToHandshake] Disable branch result optimization
This commit disables an optimization in the standard to handshake
conversion pass that would prevent the creation of Handshake-level
branches with identical inputs. This makes the implementation closer to
the legacy elastic pass, and avoids having forks being generated down
the line between branches and merges, which some passes may care about.
This optimization can be applied later during Handshake canonicalization
instead, should the user desire.