[Handshake] Fix `ChannelBufPropsAttr` text format
This commit changes the textual representation (i.e., parse/print logic)
of the `ChannelBufPropsAttr` Handshake attribute so that is is
delimiter-balanced. Previously, half-open intervals for allowed number of
slots were represented as "[<min>,inf)" which, while parsable, is in
fact disallowed by MLIR due to the unbalance of square brackets and
parenthesises. Such an interval will now be represented as "[<min>,inf]"
to satisfy MLIR's invariants.