[Handshake] Introduction of `LSQOp` (load-store queues)
This commit (finally!) introduces support for LSQs in Handshake IR.
While the operation's interface looks identical to the memory
controller's (`circt::handshake::MemoryControllerOp`), under the hood a
lot of logic is modified/added to be able to analyze its logical
control/load/store ports similarly to that of the MC. In particular,
special care was given to adding support for memory interfaces
connecting to each other, which will regularly happen between MCs and
LSQs.

This commit manages to make the vast majority of the methods on MCs and
LSQs share a single implementation and consequently avoid code/logic
duplication. This is thanks to moving some of the common logic to the
`circt::handshake::MemoryOpInterface` interface which both operations
now implement as well as some clever casting peppered in the operation's
methods.