[LSQ] Improve memory interface identification/construction
This commit fixes bugs and brings enhancements to the memory port
identification/verification logic to make it both more robust and more
flexible, especially in the situation where different memory interfaces
connect to each other (in our case, MC and LSQ). It also modifies the
LSQ builder to take a boolean indicating whether result values should be
created to eventually connect to an MC. Finally, the port identification
function used throughout the file is exposed publicly so that it can be
used from Dynamatic.