[Handshake] Do not enforce memory interface placement
This simply removes a rule in the `circt::handshake::FuncOp`
verification function that made sure that a memory interface was
connected to each memref-type function argument. In the edge case where
a memref is never accessed in the function we do not want to place a
useless memory interface for it, so we remove this rule.