[LSQ] Fix incorrect `LSQOp` parsing logic
This is a simple fix to the `handshake::LSQOp` parsing logic.
Previously, the function type at the end of the operation text format
would fail to parse in the absence of paranthesises, which is possible
whenever there is a single operand or result type. This was due to the
custom logic implemented to parse the function type, whereas the correct
way of doing this is to use a provided parsing method for function
types.