[LSQ] Fix incorrect operand naming logic
This is a small fix for the `getOperandName` method on
`handshake::LSQOp` operations, which didn't take into account the fact
that the first LSQ operand may not be the referenced memory region if
the LSQ connects to an MC (this was introduced in a more recent commit
than the LSQ, which is when the inconsistency appeared). The method now
checks whether the LSQ connects to an MC and interprets the provided
index differently based on that (offset by 1).