[Handshake] Consider function arguments valid memory control ports
This fixes a small bug in the memory port identification logic for MCs
and LSQs wherein if a control signal came directly from the start
control signal (which is a function argument) the implementation would
incorrectly declare a failure. This is in fact not a failure as it can
happen in valid dataflow circuits. The implementation is modifed to
accept such scenarios.